feat: refactor into separate function
